Decided to stray away from PDDL as formally defined.
My reasoning is two-fold:

1) I do not like the use of requirements as formally defined.  IMHO, this leaves
the language complicated and difficult to parse.

2) Requirements were intended to decouple the language from planners, as such,
   it shouldn't be on the authors of a domain/problem to know which language
   constructs imply which requirements.

EPDDL, as of now, is handling requirements on behalf of the user.  When specific
language features are used which imply a certain requirement (e.g. typing) these
requirements are automatically added to the abstract representation.

(:requirements ...) has been removed from EPDDL.